requirements.txt
aiohttp @ file:///croot/aiohttp_1694181079612/work
aiosignal @ file:///tmp/build/80754af9/aiosignal_1637843061372/work
appdirs==1.4.4
async-timeout @ file:///opt/conda/conda-bld/async-timeout_1664876359750/work
attrs @ file:///croot/attrs_1695717823297/work
beautifulsoup4 @ file:///opt/conda/conda-bld/beautifulsoup4_1650462163268/work
Bottleneck @ file:///opt/conda/conda-bld/bottleneck_1657175564434/work
brotlipy==0.7.0
certifi==2023.7.22
cffi @ file:///croot/cffi_1670423208954/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
chemicals @ file:///home/conda/feedstock_root/build_artifacts/chemicals_1694997173911/work
cryptography @ file:///croot/cryptography_1694444244250/work
discord.py @ file:///home/conda/feedstock_root/build_artifacts/discord.py_1691711949357/work
fluids @ file:///home/conda/feedstock_root/build_artifacts/fluids_1694996737421/work
frozenlist @ file:///croot/frozenlist_1670004507010/work
fuzzywuzzy @ file:///opt/conda/conda-bld/fuzzywuzzy_1650548198331/work
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645455532332/work
idna @ file:///croot/idna_1666125576474/work
lxml @ file:///opt/conda/conda-bld/lxml_1657545139709/work
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
mpmath @ file:///croot/mpmath_1690848262763/work
multidict @ file:///croot/multidict_1665674239670/work
numexpr @ file:///croot/numexpr_1683221822650/work
numpy @ file:///work/mkl/numpy_and_numpy_base_1682953417311/work
packaging @ file:///croot/packaging_1693575174725/work
pandas @ file:///croot/pandas_1692289311655/work
Pint @ file:///home/conda/feedstock_root/build_artifacts/pint_1683140320592/work
pooch @ file:///tmp/build/80754af9/pooch_1623324770023/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyOpenSSL @ file:///croot/pyopenssl_1690223430423/work
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-dotenv @ file:///home/conda/feedstock_root/build_artifacts/python-dotenv-split_1693865766170/work
python-Levenshtein @ file:///tmp/build/80754af9/python-levenshtein_1613985223962/work
pytz @ file:///croot/pytz_1695131579487/work
requests @ file:///opt/conda/conda-bld/requests_1657734628632/work
scipy==1.10.1
six @ file:///tmp/build/80754af9/six_1644875935023/work
soupsieve @ file:///croot/soupsieve_1680518478486/work
sympy @ file:///croot/sympy_1668202399572/work
tzdata @ file:///croot/python-tzdata_1690578112552/work
urllib3 @ file:///croot/urllib3_1686163155763/work
wikipedia==1.4.0
yarl @ file:///opt/conda/conda-bld/yarl_1661437085904/work
